refactor: clean up js-related configurations and remove unused headers
